Allow the avatar to be relative to the terminal
