+++ package.json
@@ -11,9 +11,7 @@
     "url": "https://github.com/geektimus/npm-card.git"
   },
   "scripts": {
-    "prepublish": "npm run build",
-    "build": "node build.js",
-    "dev": "npm run build && node ./bin/card.js",
+    "dev": "node ./bin/card.js",
     "lint": "standard",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -28,14 +26,15 @@
   "author": "Alexander Cano (http://geektimus.me)",
   "license": "MIT",
   "files": [
-    "bin/card.js",
-    "bin/output"
+    "bin/card.js"
   ],
   "devDependencies": {
+    "standard": "^12.0.1"
+  },
+  "dependencies": {
     "asciify-image": "^0.1.3",
     "boxen": "^2.1.0",
     "chalk": "^2.4.1",
-    "cli-table3": "^0.5.1",
-    "standard": "^12.0.1"
+    "cli-table3": "^0.5.1"
   }
 }
